split tests for data module
