+++ arviz/tests/test_data.py
@@ -1,12 +1,9 @@
 # pylint: disable=no-member, invalid-name, redefined-outer-name
 # pylint: disable=too-many-lines
-from collections import namedtuple, OrderedDict
+from collections import namedtuple
 import os
 from urllib.parse import urlunsplit
-
 import numpy as np
-from numpy import ma
-import pymc3 as pm
 import pytest
 import emcee
 
@@ -14,14 +11,8 @@ from arviz import (
     concat,
     convert_to_inference_data,
     convert_to_dataset,
-    from_cmdstan,
     from_dict,
-    from_pymc3,
-    from_pystan,
-    from_pyro,
-    from_emcee,
     from_netcdf,
-    from_tfp,
     to_netcdf,
     load_arviz_data,
     list_datasets,
@@ -29,29 +20,13 @@ from arviz import (
     InferenceData,
 )
 from ..data.base import generate_dims_coords, make_attrs
-from ..data.io_pystan import get_draws, get_draws_stan3  # pylint: disable=unused-import
 from ..data.datasets import REMOTE_DATASETS, LOCAL_DATASETS, RemoteFileMetadata
-from .helpers import (  # pylint: disable=unused-import
+from .helpers import (
+    chains,
     check_multiple_attrs,
-    _emcee_lnprior as emcee_lnprior,
-    _emcee_lnprob as emcee_lnprob,
-    needs_emcee3,
+    draws,
     eight_schools_params,
-    load_cached_models,
-    pystan_extract_unpermuted,
-    stan_extract_dict,
-    pystan_version,
-)
-
-
-@pytest.fixture(scope="module")
-def draws():
-    return 500
-
-
-@pytest.fixture(scope="module")
-def chains():
-    return 2
+)  # pylint: disable=unused-import
 
 
 @pytest.fixture(autouse=True)
@@ -440,96 +415,55 @@ def test_bad_inference_data():
         InferenceData(posterior=[1, 2, 3])
 
 
-class TestDictNetCDFUtils:
+class TestDataConvert:
     @pytest.fixture(scope="class")
-    def data(self, eight_schools_params, draws, chains):
-        # Data of the Eight Schools Model
-
+    def data(self, draws, chains):
         class Data:
-            _, stan_fit = load_cached_models(eight_schools_params, draws, chains)["pystan"]
-            if pystan_version() == 2:
-                stan_dict = pystan_extract_unpermuted(stan_fit)
-                obj = {}
-                for name, vals in stan_dict.items():
-                    if name not in {"y_hat", "log_lik"}:  # extra vars
-                        obj[name] = np.swapaxes(vals, 0, 1)
-            else:
-                stan_dict = stan_extract_dict(stan_fit)
-                obj = {}
-                for name, vals in stan_dict.items():
-                    if name not in {"y_hat", "log_lik"}:  # extra vars
-                        obj[name] = vals
+            # fake 8-school output
+            obs = {}
+            for key, shape in {"mu": [], "tau": [], "eta": [8], "theta": [8]}.items():
+                obs[key] = np.random.randn(draws, chains, *shape)
 
         return Data
+
+    def get_inference_data(self, data):
+        return convert_to_inference_data(
+            data.obj,
+            group="posterior",
+            coords={"school": np.arange(8)},
+            dims={"theta": ["school"], "eta": ["school"]},
+        )
 
     def check_var_names_coords_dims(self, dataset):
         assert set(dataset.data_vars) == {"mu", "tau", "eta", "theta"}
         assert set(dataset.coords) == {"chain", "draw", "school"}
 
-    def get_inference_data(self, data, eight_schools_params):
-        return convert_to_inference_data(
-            data.obj,
-            group="posterior",
-            coords={"school": np.arange(eight_schools_params["J"])},
-            dims={"theta": ["school"], "eta": ["school"]},
-        )
-
-    def test_testing_extract(self, data):
-        if pystan_version() == 2:
-            extract_func = pystan_extract_unpermuted
-            parameters = data.stan_fit.model_pars
-        else:
-            extract_func = stan_extract_dict
-            parameters = data.stan_fit.param_names
-        assert isinstance(extract_func(data.stan_fit, var_names=None), dict)
-        assert isinstance(extract_func(data.stan_fit, var_names=parameters[0]), dict)
-        assert isinstance(extract_func(data.stan_fit, var_names=parameters), dict)
-        assert isinstance(
-            extract_func(data.stan_fit, var_names=[parameters[0], parameters[0]]), dict
-        )
-
-    def test_convert_to_inference_data(self, data, eight_schools_params):
-        inference_data = self.get_inference_data(data, eight_schools_params)
+    def test_convert_to_inference_data(self, data):
+        inference_data = self.get_inference_data(data)
         assert hasattr(inference_data, "posterior")
         self.check_var_names_coords_dims(inference_data.posterior)
 
-    def test_convert_to_dataset(self, eight_schools_params, draws, chains, data):
+    def test_convert_to_dataset(self, draws, chains, data):
         dataset = convert_to_dataset(
             data.obj,
             group="posterior",
-            coords={"school": np.arange(eight_schools_params["J"])},
+            coords={"school": np.arange(8)},
             dims={"theta": ["school"], "eta": ["school"]},
         )
         assert dataset.draw.shape == (draws,)
         assert dataset.chain.shape == (chains,)
-        assert dataset.school.shape == (eight_schools_params["J"],)
-        assert dataset.theta.shape == (chains, draws, eight_schools_params["J"])
-
-    def test_from_dict_warning(self):
-        bad_posterior_dict = {"log_likelihood": np.ones((5, 1000, 2))}
-        with pytest.warns(SyntaxWarning):
-            from_dict(posterior=bad_posterior_dict)
+        assert dataset.school.shape == (8,)
+        assert dataset.theta.shape == (chains, draws, 8)
 
 
-class TestDictIONetCDFUtils:
+class TestDataDict:
     @pytest.fixture(scope="class")
-    def data(self, eight_schools_params, draws, chains):
-        # Data of the Eight Schools Model
-
+    def data(self, draws, chains):
         class Data:
-            _, stan_fit = load_cached_models(eight_schools_params, draws, chains)["pystan"]
-            if pystan_version() == 2:
-                stan_dict = pystan_extract_unpermuted(stan_fit)
-                obj = {}
-                for name, vals in stan_dict.items():
-                    if name not in {"y_hat", "log_lik"}:  # extra vars
-                        obj[name] = np.swapaxes(vals, 0, 1)
-            else:
-                stan_dict = stan_extract_dict(stan_fit)
-                obj = {}
-                for name, vals in stan_dict.items():
-                    if name not in {"y_hat", "log_lik"}:  # extra vars
-                        obj[name] = vals
+            # fake 8-school output
+            obs = {}
+            for key, shape in {"mu": [], "tau": [], "eta": [8], "theta": [8]}.items():
+                obs[key] = np.random.randn(draws, chains, *shape)
 
         return Data
 
@@ -546,7 +480,7 @@ class TestDictIONetCDFUtils:
             prior_predictive=data.obj,
             sample_stats_prior=data.obj,
             observed_data=eight_schools_params,
-            coords={"school": np.arange(eight_schools_params["J"])},
+            coords={"school": np.arange(8)},
             dims={"theta": ["school"], "eta": ["school"]},
         )
 
@@ -603,148 +537,33 @@ class TestDictIONetCDFUtils:
         with pytest.raises(TypeError):
             from_dict(observed_data=x)
 
-
-class TestEmceeNetCDFUtils:
-    arg_list = [
-        ({}, {"posterior": ["var_0", "var_1", "var_7"], "observed_data": ["arg_0", "arg_1"]}),
-        (
-            {"var_names": ["mu", "tau", "eta"], "slices": [0, 1, slice(2, None)]},
-            {"posterior": ["mu", "tau", "eta"], "observed_data": ["arg_0", "arg_1"]},
-        ),
-        (
-            {
-                "blob_names": ["log_likelihood", "y"],
-                "blob_groups": ["sample_stats", "posterior_predictive"],
-            },
-            {
-                "posterior": ["var_0", "var_1", "var_7"],
-                "observed_data": ["arg_0", "arg_1"],
-                "sample_stats": ["log_likelihood"],
-                "posterior_predictive": ["y"],
-            },
-        ),
-        (
-            {
-                "blob_names": ["log_likelihood", "y"],
-                "dims": {"eta": ["school"], "log_likelihood": ["school"], "y": ["school"]},
-                "var_names": ["mu", "tau", "eta"],
-                "slices": [0, 1, slice(2, None)],
-                "arg_names": ["y", "sigma"],
-                "coords": {"school": range(8)},
-            },
-            {
-                "posterior": ["mu", "tau", "eta"],
-                "observed_data": ["y", "sigma"],
-                "sample_stats": ["log_likelihood", "y"],
-            },
-        ),
-    ]
-
-    @pytest.fixture(scope="class")
-    def data(self, draws, chains):
-        class Data:
-            # chains are not used
-            # emcee uses lots of walkers
-            obj = load_cached_models(eight_schools_params, draws, chains)["emcee"]
-
-        return Data
-
-    def get_inference_data_reader(self, **kwargs):
-        from emcee import backends  # pylint: disable=no-name-in-module
-
-        here = os.path.dirname(os.path.abspath(__file__))
-        data_directory = os.path.join(here, "saved_models")
-        filepath = os.path.join(data_directory, "reader_testfile.h5")
-        assert os.path.exists(filepath)
-        assert os.path.getsize(filepath)
-        reader = backends.HDFBackend(filepath, read_only=True)
-        return from_emcee(reader, **kwargs)
-
-    @pytest.mark.parametrize("test_args", arg_list)
-    def test_inference_data(self, data, test_args):
-        kwargs, test_dict = test_args
-        inference_data = from_emcee(data.obj, **kwargs)
-        fails = check_multiple_attrs(test_dict, inference_data)
-        assert not fails
-
-    @needs_emcee3
-    @pytest.mark.parametrize("test_args", arg_list)
-    def test_inference_data_reader(self, test_args):
-        kwargs, test_dict = test_args
-        kwargs = {k: i for k, i in kwargs.items() if k != "arg_names"}
-        inference_data = self.get_inference_data_reader(**kwargs)
-        test_dict.pop("observed_data")
-        fails = check_multiple_attrs(test_dict, inference_data)
-        assert not fails
-
-    def test_verify_var_names(self, data):
-        with pytest.raises(ValueError):
-            from_emcee(data.obj, var_names=["not", "enough"])
-
-    def test_verify_arg_names(self, data):
-        with pytest.raises(ValueError):
-            from_emcee(data.obj, arg_names=["not enough"])
-
-    @pytest.mark.parametrize("slices", [[0, 0, slice(2, None)], [0, 1, slice(1, None)]])
-    def test_slices_warning(self, data, slices):
+    def test_from_dict_warning(self):
+        bad_posterior_dict = {"log_likelihood": np.ones((5, 1000, 2))}
         with pytest.warns(SyntaxWarning):
-            from_emcee(data.obj, slices=slices)
-
-    def test_no_blobs_error(self):
-        sampler = emcee.EnsembleSampler(6, 1, lambda x: -x ** 2)
-        sampler.run_mcmc(np.random.normal(size=(6, 1)), 20)
-        with pytest.raises(ValueError):
-            from_emcee(sampler, blob_names=["inexistent"])
-
-    def test_peculiar_blobs(self, data):
-        sampler = emcee.EnsembleSampler(6, 1, lambda x: (-x ** 2, (np.random.normal(x), 3)))
-        sampler.run_mcmc(np.random.normal(size=(6, 1)), 20)
-        inference_data = from_emcee(sampler, blob_names=["normal", "threes"])
-        fails = check_multiple_attrs({"sample_stats": ["normal", "threes"]}, inference_data)
-        assert not fails
-        inference_data = from_emcee(data.obj, blob_names=["mix"])
-        fails = check_multiple_attrs({"sample_stats": ["mix"]}, inference_data)
-        assert not fails
-
-    @pytest.mark.parametrize(
-        "blob_args",
-        [
-            (ValueError, ["a", "b"], ["prior"]),
-            (ValueError, ["too", "many", "names"], None),
-            (SyntaxError, ["a", "b"], ["posterior", "observed_data"]),
-        ],
-    )
-    def test_bad_blobs(self, data, blob_args):
-        error, names, groups = blob_args
-        with pytest.raises(error):
-            from_emcee(data.obj, blob_names=names, blob_groups=groups)
-
-    def test_ln_funcs_for_infinity(self):
-        # after dropping Python 3.5 support use underscore 1_000_000
-        ary = np.ones(10)
-        ary[1] = -1
-        assert np.isinf(emcee_lnprior(ary))
-        assert np.isinf(emcee_lnprob(ary, ary[2:], ary[2:])[0])
+            from_dict(posterior=bad_posterior_dict)
 
 
-class TestIONetCDFUtils:
+class TestDataNetCDF:
     @pytest.fixture(scope="class")
     def data(self, draws, chains):
         class Data:
-            model, obj = load_cached_models(eight_schools_params, draws, chains)["pymc3"]
+            # fake 8-school output
+            obs = {}
+            for key, shape in {"mu": [], "tau": [], "eta": [8], "theta": [8]}.items():
+                obs[key] = np.random.randn(draws, chains, *shape)
 
         return Data
 
-    def get_inference_data(self, data, eight_schools_params):  # pylint: disable=W0613
-        with data.model:
-            prior = pm.sample_prior_predictive()
-            posterior_predictive = pm.sample_posterior_predictive(data.obj)
-
-        return from_pymc3(
-            trace=data.obj,
-            prior=prior,
-            posterior_predictive=posterior_predictive,
-            coords={"school": np.arange(eight_schools_params["J"])},
+    def get_inference_data(self, data, eight_schools_params):
+        return from_dict(
+            posterior=data.obj,
+            posterior_predictive=data.obj,
+            sample_stats=data.obj,
+            prior=data.obj,
+            prior_predictive=data.obj,
+            sample_stats_prior=data.obj,
+            observed_data=eight_schools_params,
+            coords={"school": np.arange(8)},
             dims={"theta": ["school"], "eta": ["school"]},
         )
 
@@ -793,733 +612,3 @@ class TestIONetCDFUtils:
         assert os.path.exists(filepath)
         os.remove(filepath)
         assert not os.path.exists(filepath)
-
-
-class TestPyMC3NetCDFUtils:
-    @pytest.fixture(scope="class")
-    def data(self, draws, chains):
-        class Data:
-            model, obj = load_cached_models(eight_schools_params, draws, chains)["pymc3"]
-
-        return Data
-
-    def get_inference_data(self, data, eight_schools_params):
-        with data.model:
-            prior = pm.sample_prior_predictive()
-            posterior_predictive = pm.sample_posterior_predictive(data.obj)
-
-        return from_pymc3(
-            trace=data.obj,
-            prior=prior,
-            posterior_predictive=posterior_predictive,
-            coords={"school": np.arange(eight_schools_params["J"])},
-            dims={"theta": ["school"], "eta": ["school"]},
-        )
-
-    def test_posterior(self, data, eight_schools_params):
-        inference_data = self.get_inference_data(data, eight_schools_params)
-        assert hasattr(inference_data, "posterior")
-
-    def test_sampler_stats(self, data, eight_schools_params):
-        inference_data = self.get_inference_data(data, eight_schools_params)
-        assert hasattr(inference_data, "sample_stats")
-
-    def test_posterior_predictive(self, data, eight_schools_params):
-        inference_data = self.get_inference_data(data, eight_schools_params)
-        assert hasattr(inference_data, "posterior_predictive")
-
-    def test_prior(self, data, eight_schools_params):
-        inference_data = self.get_inference_data(data, eight_schools_params)
-        assert hasattr(inference_data, "prior")
-
-    def test_missing_data_model(self):
-        # source pymc3/pymc3/tests/test_missing.py
-        data = ma.masked_values([1, 2, -1, 4, -1], value=-1)
-        model = pm.Model()
-        with model:
-            x = pm.Normal("x", 1, 1)
-            pm.Normal("y", x, 1, observed=data)
-            trace = pm.sample(100, chains=2)
-
-        # make sure that data is really missing
-        y_missing, = model.missing_values
-        assert y_missing.tag.test_value.shape == (2,)
-        inference_data = from_pymc3(trace=trace)
-        test_dict = {"posterior": ["x"], "observed_data": ["y"], "sample_stats": ["log_likelihood"]}
-        fails = check_multiple_attrs(test_dict, inference_data)
-        assert not fails
-
-    def test_multiple_observed_rv(self):
-        y1_data = np.random.randn(10)
-        y2_data = np.random.randn(100)
-        with pm.Model():
-            x = pm.Normal("x", 1, 1)
-            pm.Normal("y1", x, 1, observed=y1_data)
-            pm.Normal("y2", x, 1, observed=y2_data)
-            trace = pm.sample(100, chains=2)
-        inference_data = from_pymc3(trace=trace)
-        test_dict = {"posterior": ["x"], "observed_data": ["y1", "y2"], "sample_stats": ["lp"]}
-        fails = check_multiple_attrs(test_dict, inference_data)
-        assert not fails
-        assert not hasattr(inference_data.sample_stats, "log_likelihood")
-
-
-class TestPyroNetCDFUtils:
-    @pytest.fixture(scope="class")
-    def data(self, draws, chains):
-        class Data:
-            obj = load_cached_models(eight_schools_params, draws, chains)["pyro"]
-
-        return Data
-
-    def get_inference_data(self, data):
-        return from_pyro(posterior=data.obj)
-
-    def test_inference_data(self, data):
-        inference_data = self.get_inference_data(data)
-        assert hasattr(inference_data, "posterior")
-
-
-class TestPyStanNetCDFUtils:
-    @pytest.fixture(scope="class")
-    def data(self, draws, chains):
-        class Data:
-            model, obj = load_cached_models(eight_schools_params, draws, chains)["pystan"]
-
-        return Data
-
-    def get_inference_data(self, data, eight_schools_params):
-        """vars as str."""
-        return from_pystan(
-            posterior=data.obj,
-            posterior_predictive="y_hat",
-            prior=data.obj,
-            prior_predictive="y_hat",
-            observed_data="y",
-            log_likelihood="log_lik",
-            coords={"school": np.arange(eight_schools_params["J"])},
-            dims={
-                "theta": ["school"],
-                "y": ["school"],
-                "log_lik": ["school"],
-                "y_hat": ["school"],
-                "eta": ["school"],
-            },
-            posterior_model=data.model,
-            prior_model=data.model,
-        )
-
-    def get_inference_data2(self, data, eight_schools_params):
-        """vars as lists."""
-        return from_pystan(
-            posterior=data.obj,
-            posterior_predictive=["y_hat"],
-            prior=data.obj,
-            prior_predictive=["y_hat"],
-            observed_data="y",
-            log_likelihood="log_lik",
-            coords={
-                "school": np.arange(eight_schools_params["J"]),
-                "log_likelihood_dim": np.arange(eight_schools_params["J"]),
-            },
-            dims={
-                "theta": ["school"],
-                "y": ["school"],
-                "y_hat": ["school"],
-                "eta": ["school"],
-                "log_lik": ["log_likelihood_dim"],
-            },
-            posterior_model=data.model,
-            prior_model=data.model,
-        )
-
-    def get_inference_data3(self, data, eight_schools_params):
-        """multiple vars as lists."""
-        return from_pystan(
-            posterior=data.obj,
-            posterior_predictive=["y_hat", "log_lik"],
-            prior=data.obj,
-            prior_predictive=["y_hat", "log_lik"],
-            observed_data="y",
-            coords={"school": np.arange(eight_schools_params["J"])},
-            dims={"theta": ["school"], "y": ["school"], "y_hat": ["school"], "eta": ["school"]},
-            posterior_model=data.model,
-            prior_model=data.model,
-        )
-
-    def get_inference_data4(self, data):
-        """multiple vars as lists."""
-        return from_pystan(
-            posterior=data.obj,
-            posterior_predictive=None,
-            prior=data.obj,
-            prior_predictive=None,
-            observed_data="y",
-            coords=None,
-            dims=None,
-            posterior_model=data.model,
-            prior_model=data.model,
-        )
-
-    def test_sampler_stats(self, data, eight_schools_params):
-        inference_data = self.get_inference_data(data, eight_schools_params)
-        test_dict = {"sample_stats": ["lp", "diverging"]}
-        fails = check_multiple_attrs(test_dict, inference_data)
-        assert not fails
-
-    def test_inference_data(self, data, eight_schools_params):
-        inference_data1 = self.get_inference_data(data, eight_schools_params)
-        inference_data2 = self.get_inference_data2(data, eight_schools_params)
-        inference_data3 = self.get_inference_data3(data, eight_schools_params)
-        inference_data4 = self.get_inference_data4(data)
-        # inference_data 1
-        test_dict = {
-            "posterior": ["theta"],
-            "observed_data": ["y"],
-            "sample_stats": ["log_likelihood"],
-            "prior": ["theta"],
-        }
-        fails = check_multiple_attrs(test_dict, inference_data1)
-        assert not fails
-        # inference_data 2
-        test_dict = {
-            "posterior_predictive": ["y_hat"],
-            "observed_data": ["y"],
-            "sample_stats_prior": ["lp"],
-            "sample_stats": ["lp"],
-            "prior_predictive": ["y_hat"],
-        }
-        fails = check_multiple_attrs(test_dict, inference_data2)
-        assert not fails
-        # inference_data 3
-        test_dict = {
-            "posterior_predictive": ["y_hat"],
-            "observed_data": ["y"],
-            "sample_stats_prior": ["lp"],
-            "sample_stats": ["lp"],
-            "prior_predictive": ["y_hat"],
-        }
-        fails = check_multiple_attrs(test_dict, inference_data3)
-        assert not fails
-        # inference_data 4
-        test_dict = {"posterior": ["theta"], "prior": ["theta"]}
-        fails = check_multiple_attrs(test_dict, inference_data4)
-        assert not fails
-
-    def test_invalid_fit(self, data):
-        if pystan_version() == 2:
-            model = data.model
-            model_data = {
-                "J": 8,
-                "y": np.array([28.0, 8.0, -3.0, 7.0, -1.0, 1.0, 18.0, 12.0]),
-                "sigma": np.array([15.0, 10.0, 16.0, 11.0, 9.0, 11.0, 10.0, 18.0]),
-            }
-            fit_test_grad = model.sampling(
-                data=model_data, test_grad=True, check_hmc_diagnostics=False
-            )
-            with pytest.raises(AttributeError):
-                _ = from_pystan(posterior=fit_test_grad)
-            fit = model.sampling(data=model_data, iter=100, chains=1, check_hmc_diagnostics=False)
-            del fit.sim["samples"]
-            with pytest.raises(AttributeError):
-                _ = from_pystan(posterior=fit)
-
-    def test_empty_parameter(self):
-        if pystan_version() == 2:
-            model_code = """
-                parameters {
-                    real y;
-                    vector[3] x;
-                    vector[0] a;
-                    vector[2] z;
-                }
-                model {
-                    y ~ normal(0,1);
-                }
-            """
-            from pystan import StanModel
-
-            model = StanModel(model_code=model_code)
-            fit = model.sampling(iter=10, chains=2, check_hmc_diagnostics=False)
-            posterior = from_pystan(posterior=fit)
-            test_dict = {"posterior": ["y", "x", "z"], "sample_stats": ["lp"]}
-            fails = check_multiple_attrs(test_dict, posterior)
-            assert not fails
-
-    def test_get_draws(self, data):
-        fit = data.obj
-        if pystan_version() == 2:
-            draws = get_draws(fit, variables=["theta", "theta"])
-            assert draws.get("theta") is not None
-        else:
-            draws = get_draws_stan3(fit, variables=["theta", "theta"])
-            assert draws.get("theta") is not None
-
-    @pytest.mark.skipif(pystan_version() != 2, reason="PyStan 2.x required")
-    def test_index_order(self, data, eight_schools_params):
-        """Test 0-indexed data."""
-        import pystan
-
-        fit = data.model.sampling(data=eight_schools_params)
-        if pystan.__version__ >= "2.18":
-            # make 1-indexed to 0-indexed
-            for holder in fit.sim["samples"]:
-                new_chains = OrderedDict()
-                for i, (key, values) in enumerate(holder.chains.items()):
-                    if "[" in key:
-                        name, *shape = key.replace("]", "").split("[")
-                        shape = [str(int(item) - 1) for items in shape for item in items.split(",")]
-                        key = name + "[{}]".format(",".join(shape))
-                    new_chains[key] = np.full_like(values, fill_value=float(i))
-                setattr(holder, "chains", new_chains)
-            fit.sim["fnames_oi"] = list(fit.sim["samples"][0].chains.keys())
-        idata = from_pystan(posterior=fit)
-        assert idata is not None
-        for j, fpar in enumerate(fit.sim["fnames_oi"]):
-            if fpar == "lp__":
-                continue
-            par, *shape = fpar.replace("]", "").split("[")
-            assert hasattr(idata.posterior, par)
-            if shape:
-                shape = [slice(None), slice(None)] + list(map(int, shape))
-                assert idata.posterior[par][tuple(shape)].values.mean() == float(j)
-            else:
-                assert idata.posterior[par].values.mean() == float(j)
-
-
-class TestTfpNetCDFUtils:
-    @pytest.fixture(scope="class")
-    def data(self, draws, chains):
-        class Data:
-            # Returns result of from_tfp
-            model, obj = load_cached_models(eight_schools_params, draws, chains)[
-                "tensorflow_probability"
-            ]
-
-        return Data
-
-    def get_inference_data(self, data, eight_schools_params):
-        """Normal read with observed and var_names."""
-        inference_data = from_tfp(
-            data.obj,
-            var_names=["mu", "tau", "eta"],
-            model_fn=lambda: data.model(
-                eight_schools_params["J"], eight_schools_params["sigma"].astype(np.float32)
-            ),
-            observed=eight_schools_params["y"].astype(np.float32),
-        )
-        return inference_data
-
-    def get_inference_data2(self, data):
-        """Fit only."""
-        inference_data = from_tfp(data.obj)
-        return inference_data
-
-    def get_inference_data3(self, data, eight_schools_params):
-        """Read with observed Tensor var_names and dims."""
-        import tensorflow as tf
-
-        inference_data = from_tfp(
-            data.obj,
-            var_names=["mu", "tau", "eta"],
-            model_fn=lambda: data.model(
-                eight_schools_params["J"], eight_schools_params["sigma"].astype(np.float32)
-            ),
-            posterior_predictive_samples=100,
-            posterior_predictive_size=3,
-            observed=tf.convert_to_tensor(
-                np.vstack(
-                    (
-                        eight_schools_params["y"],
-                        eight_schools_params["y"],
-                        eight_schools_params["y"],
-                    )
-                ).astype(np.float32),
-                np.float32,
-            ),
-            coords={"school": np.arange(eight_schools_params["J"])},
-            dims={"eta": ["school"], "obs": ["size_dim", "school"]},
-        )
-        return inference_data
-
-    def get_inference_data4(self, data, eight_schools_params):
-        """Test setter."""
-        inference_data = from_tfp(
-            data.obj + [np.ones_like(data.obj[0]).astype(np.float32)],
-            var_names=["mu", "tau", "eta", "avg_effect"],
-            model_fn=lambda: data.model(
-                eight_schools_params["J"], eight_schools_params["sigma"].astype(np.float32)
-            ),
-            observed=eight_schools_params["y"].astype(np.float32),
-        )
-        return inference_data
-
-    def test_inference_data(self, data, eight_schools_params):
-        inference_data = self.get_inference_data(data, eight_schools_params)
-        test_dict = {
-            "posterior": ["mu", "tau", "eta"],
-            "observed_data": ["obs"],
-            "posterior_predictive": ["obs"],
-        }
-        fails = check_multiple_attrs(test_dict, inference_data)
-        assert not fails
-
-    def test_inference_data2(self, data):
-        inference_data = self.get_inference_data2(data)
-        assert hasattr(inference_data, "posterior")
-
-    def test_inference_data3(self, data, eight_schools_params):
-        inference_data = self.get_inference_data3(data, eight_schools_params)
-        test_dict = {
-            "posterior": ["mu", "tau", "eta"],
-            "observed_data": ["obs"],
-            "posterior_predictive": ["obs"],
-        }
-        fails = check_multiple_attrs(test_dict, inference_data)
-        assert not fails
-
-    def test_inference_data4(self, data, eight_schools_params):
-        inference_data = self.get_inference_data4(data, eight_schools_params)
-        test_dict = {
-            "posterior": ["mu", "tau", "eta", "avg_effect"],
-            "observed_data": ["obs"],
-            "posterior_predictive": ["obs"],
-        }
-        fails = check_multiple_attrs(test_dict, inference_data)
-        assert not fails
-
-
-class TestCmdStanNetCDFUtils:
-    @pytest.fixture(scope="session")
-    def data_directory(self):
-        here = os.path.dirname(os.path.abspath(__file__))
-        data_directory = os.path.join(here, "saved_models")
-        return data_directory
-
-    @pytest.fixture(scope="class")
-    def paths(self, data_directory):
-        paths = {
-            "no_warmup": [
-                os.path.join(data_directory, "cmdstan/output_no_warmup1.csv"),
-                os.path.join(data_directory, "cmdstan/output_no_warmup2.csv"),
-                os.path.join(data_directory, "cmdstan/output_no_warmup3.csv"),
-                os.path.join(data_directory, "cmdstan/output_no_warmup4.csv"),
-            ],
-            "warmup": [
-                os.path.join(data_directory, "cmdstan/output_warmup1.csv"),
-                os.path.join(data_directory, "cmdstan/output_warmup2.csv"),
-                os.path.join(data_directory, "cmdstan/output_warmup3.csv"),
-                os.path.join(data_directory, "cmdstan/output_warmup4.csv"),
-            ],
-            "no_warmup_glob": os.path.join(data_directory, "cmdstan/output_no_warmup[0-9].csv"),
-            "warmup_glob": os.path.join(data_directory, "cmdstan/output_warmup[0-9].csv"),
-            "combined_no_warmup": [
-                os.path.join(data_directory, "cmdstan/combined_output_no_warmup.csv")
-            ],
-            "combined_warmup": [os.path.join(data_directory, "cmdstan/combined_output_warmup.csv")],
-            "combined_no_warmup_glob": os.path.join(
-                data_directory, "cmdstan/combined_output_no_warmup.csv"
-            ),
-            "combined_warmup_glob": os.path.join(
-                data_directory, "cmdstan/combined_output_warmup.csv"
-            ),
-            "eight_schools_glob": os.path.join(
-                data_directory, "cmdstan/eight_schools_output[0-9].csv"
-            ),
-            "eight_schools": [
-                os.path.join(data_directory, "cmdstan/eight_schools_output1.csv"),
-                os.path.join(data_directory, "cmdstan/eight_schools_output2.csv"),
-                os.path.join(data_directory, "cmdstan/eight_schools_output3.csv"),
-                os.path.join(data_directory, "cmdstan/eight_schools_output4.csv"),
-            ],
-            "missing_files": [
-                os.path.join(data_directory, "cmdstan/combined_missing_config.csv"),
-                os.path.join(data_directory, "cmdstan/combined_missing_adaptation.csv"),
-                os.path.join(data_directory, "cmdstan/combined_missing_timing1.csv"),
-                os.path.join(data_directory, "cmdstan/combined_missing_timing2.csv"),
-            ],
-        }
-        return paths
-
-    @pytest.fixture(scope="class")
-    def observed_data_paths(self, data_directory):
-        observed_data_paths = [
-            os.path.join(data_directory, "cmdstan/eight_schools.data.R"),
-            os.path.join(data_directory, "cmdstan/example_stan.data.R"),
-        ]
-
-        return observed_data_paths
-
-    def get_inference_data(self, posterior, **kwargs):
-        return from_cmdstan(posterior=posterior, **kwargs)
-
-    def test_sample_stats(self, paths):
-        for key, path in paths.items():
-            if "missing" in key:
-                continue
-            inference_data = self.get_inference_data(path)
-            assert hasattr(inference_data, "sample_stats")
-
-    def test_inference_data_shapes(self, paths):
-        """Assert that shapes are transformed correctly"""
-        for key, path in paths.items():
-            if "eight" in key or "missing" in key:
-                continue
-            inference_data = self.get_inference_data(path)
-            test_dict = {"posterior": ["x", "y", "Z"]}
-            fails = check_multiple_attrs(test_dict, inference_data)
-            assert not fails
-            assert inference_data.posterior["y"].shape == (4, 100)
-            assert inference_data.posterior["x"].shape == (4, 100, 3)
-            assert inference_data.posterior["Z"].shape == (4, 100, 4, 6)
-            dims = ["chain", "draw"]
-            y_mean_true = 0
-            y_mean = inference_data.posterior["y"].mean(dim=dims)
-            assert np.isclose(y_mean, y_mean_true, atol=1e-1)
-            x_mean_true = np.array([1, 2, 3])
-            x_mean = inference_data.posterior["x"].mean(dim=dims)
-            assert np.isclose(x_mean, x_mean_true, atol=1e-1).all()
-            Z_mean_true = np.array([1, 2, 3, 4])
-            Z_mean = inference_data.posterior["Z"].mean(dim=dims).mean(axis=1)
-            assert np.isclose(Z_mean, Z_mean_true, atol=7e-1).all()
-
-    def test_inference_data_input_types1(self, paths, observed_data_paths):
-        """Check input types
-
-            posterior --> str, list of str
-            prior --> str, list of str
-            posterior_predictive --> str, variable in posterior
-            observed_data --> Rdump format
-            observed_data_var --> str, variable
-            log_likelihood --> str
-            coords --> one to many
-            dims --> one to many
-        """
-        for key, path in paths.items():
-            if "eight" not in key:
-                continue
-            inference_data = self.get_inference_data(
-                posterior=path,
-                posterior_predictive="y_hat",
-                prior=path,
-                prior_predictive="y_hat",
-                observed_data=observed_data_paths[0],
-                observed_data_var="y",
-                log_likelihood="log_lik",
-                coords={"school": np.arange(8)},
-                dims={
-                    "theta": ["school"],
-                    "y": ["school"],
-                    "log_lik": ["school"],
-                    "y_hat": ["school"],
-                    "eta": ["school"],
-                },
-            )
-            test_dict = {
-                "posterior": ["mu", "tau", "theta_tilde", "theta"],
-                "prior": ["mu", "tau", "theta_tilde", "theta"],
-                "prior_predictive": ["y_hat"],
-                "sample_stats": ["log_likelihood"],
-                "observed_data": ["y"],
-                "posterior_predictive": ["y_hat"],
-            }
-            fails = check_multiple_attrs(test_dict, inference_data)
-            assert not fails
-
-    def test_inference_data_input_types2(self, paths, observed_data_paths):
-        """Check input types (change, see earlier)
-
-            posterior_predictive --> List[str], variable in posterior
-            observed_data_var --> List[str], variable
-        """
-        for key, path in paths.items():
-            if "eight" not in key:
-                continue
-            inference_data = self.get_inference_data(
-                posterior=path,
-                posterior_predictive=["y_hat"],
-                prior=path,
-                prior_predictive=["y_hat"],
-                observed_data=observed_data_paths[0],
-                observed_data_var=["y"],
-                log_likelihood="log_lik",
-                coords={"school": np.arange(8)},
-                dims={
-                    "theta": ["school"],
-                    "y": ["school"],
-                    "log_lik": ["school"],
-                    "y_hat": ["school"],
-                    "eta": ["school"],
-                },
-            )
-            test_dict = {
-                "posterior": ["mu", "tau", "theta_tilde", "theta"],
-                "prior": ["mu", "tau", "theta_tilde", "theta"],
-                "prior_predictive": ["y_hat"],
-                "sample_stats": ["log_likelihood"],
-                "observed_data": ["y"],
-                "posterior_predictive": ["y_hat"],
-            }
-            fails = check_multiple_attrs(test_dict, inference_data)
-            assert not fails
-
-    def test_inference_data_input_types3(self, paths, observed_data_paths):
-        """Check input types (change, see earlier)
-
-            posterior_predictive --> str, csv file
-            coords --> one to many + one to one (default dim)
-            dims --> one to many
-        """
-        for key, path in paths.items():
-            if "eight" not in key:
-                continue
-            post_pred = paths["eight_schools_glob"]
-            inference_data = self.get_inference_data(
-                posterior=path,
-                posterior_predictive=post_pred,
-                prior=path,
-                prior_predictive=post_pred,
-                observed_data=observed_data_paths[0],
-                observed_data_var=["y"],
-                log_likelihood="log_lik",
-                coords={"school": np.arange(8), "log_lik_dim_0": np.arange(8)},
-                dims={"theta": ["school"], "y": ["school"], "y_hat": ["school"], "eta": ["school"]},
-            )
-            test_dict = {
-                "posterior": ["mu", "tau", "theta_tilde", "theta"],
-                "prior": ["mu", "tau", "theta_tilde", "theta"],
-                "prior_predictive": ["y_hat"],
-                "sample_stats": ["log_likelihood"],
-                "observed_data": ["y"],
-                "posterior_predictive": ["y_hat"],
-            }
-            fails = check_multiple_attrs(test_dict, inference_data)
-            assert not fails
-
-    def test_inference_data_input_types4(self, paths):
-        """Check input types (change, see earlier)
-
-            coords --> one to many + one to one (non-default dim)
-            dims --> one to many + one to one
-        """
-
-        path = paths["combined_no_warmup"]
-        for path in [path, path[0]]:
-            inference_data = self.get_inference_data(
-                posterior=path,
-                posterior_predictive=path,
-                prior=path,
-                prior_predictive=path,
-                observed_data=None,
-                observed_data_var=None,
-                coords={"rand": np.arange(3)},
-                dims={"x": ["rand"]},
-            )
-            test_dict = {
-                "posterior": ["x", "y", "Z"],
-                "prior": ["x", "y", "Z"],
-                "prior_predictive": ["x", "y", "Z"],
-                "sample_stats": ["lp"],
-                "sample_stats_prior": ["lp"],
-                "posterior_predictive": ["x", "y", "Z"],
-            }
-            fails = check_multiple_attrs(test_dict, inference_data)
-            assert not fails
-
-    def test_inference_data_input_types5(self, paths, observed_data_paths):
-        """Check input types (change, see earlier)
-
-            posterior_predictive is None
-            prior_predictive is None
-        """
-        for key, path in paths.items():
-            if "eight" not in key:
-                continue
-            inference_data = self.get_inference_data(
-                posterior=path,
-                posterior_predictive=None,
-                prior=path,
-                prior_predictive=None,
-                observed_data=observed_data_paths[0],
-                observed_data_var=["y"],
-                log_likelihood=["log_lik"],
-                coords={"school": np.arange(8), "log_lik_dim": np.arange(8)},
-                dims={
-                    "theta": ["school"],
-                    "y": ["school"],
-                    "log_lik": ["log_lik_dim"],
-                    "y_hat": ["school"],
-                    "eta": ["school"],
-                },
-            )
-            test_dict = {
-                "posterior": ["mu", "tau", "theta_tilde", "theta"],
-                "prior": ["mu", "tau", "theta_tilde", "theta"],
-                "sample_stats": ["log_likelihood"],
-                "observed_data": ["y"],
-                "sample_stats_prior": ["lp"],
-            }
-            fails = check_multiple_attrs(test_dict, inference_data)
-            assert not fails
-
-    def test_inference_data_bad_csv(self, paths):
-        """Check ValueError for csv with missing headers"""
-        for key, _paths in paths.items():
-            if "missing" not in key:
-                continue
-            for path in _paths:
-                with pytest.raises(ValueError):
-                    self.get_inference_data(posterior=path)
-
-    def test_inference_data_observed_data1(self, observed_data_paths):
-        """Read Rdump, check shapes are correct
-
-            All variables
-        """
-        path = observed_data_paths[1]
-        inference_data = self.get_inference_data(posterior=None, observed_data=path)
-        assert hasattr(inference_data, "observed_data")
-        assert len(inference_data.observed_data.data_vars) == 3
-        assert inference_data.observed_data["x"].shape == (1,)
-        assert inference_data.observed_data["y"].shape == (3,)
-        assert inference_data.observed_data["Z"].shape == (4, 5)
-
-    def test_inference_data_observed_data2(self, observed_data_paths):
-        """Read Rdump, check shapes are correct
-
-            One variable as str
-        """
-        path = observed_data_paths[1]
-        inference_data = self.get_inference_data(
-            posterior=None, observed_data=path, observed_data_var="x"
-        )
-        assert hasattr(inference_data, "observed_data")
-        assert len(inference_data.observed_data.data_vars) == 1
-        assert inference_data.observed_data["x"].shape == (1,)
-
-    def test_inference_data_observed_data3(self, observed_data_paths):
-        """Read Rdump, check shapes are correct
-
-            One variable as a list
-        """
-        path = observed_data_paths[1]
-        inference_data = self.get_inference_data(
-            posterior=None, observed_data=path, observed_data_var=["x"]
-        )
-        assert hasattr(inference_data, "observed_data")
-        assert len(inference_data.observed_data.data_vars) == 1
-        assert inference_data.observed_data["x"].shape == (1,)
-
-    def test_inference_data_observed_data4(self, observed_data_paths):
-        """Read Rdump, check shapes are correct
-
-            Many variables as list
-        """
-        path = observed_data_paths[1]
-        inference_data = self.get_inference_data(
-            posterior=None, observed_data=path, observed_data_var=["y", "Z"]
-        )
-        assert hasattr(inference_data, "observed_data")
-        assert len(inference_data.observed_data.data_vars) == 2
-        assert inference_data.observed_data["y"].shape == (3,)
-        assert inference_data.observed_data["Z"].shape == (4, 5)
